+++ README.md
@@ -36,7 +36,7 @@
 
 ### 迭代记录
 
-#### 1.0.3-rc.14
+#### 1.0.3-rc.15
 * 新增全局错误处理
 
 #### 1.0.3-rc.12
