+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "astroboy.ts",
-  "version": "1.0.3-rc.14",
+  "version": "1.0.3-rc.15",
   "description": "基于[astroboy]的DI版本，TypeScript3.2信仰加成",
   "main": "./dist/index.js",
   "scripts": {
